+++ x/evm/keeper/grpc_query.go
@@ -265,8 +265,7 @@ func (k Keeper) EstimateGas(c context.Context, req *types.EthCallRequest) (*type
 }
 
 // EstimateGasInternal returns the gas estimation for the corresponding request.
-// This function is called from the RPC client (eth_estimateGas) and internally
-// by the CallEVMWithData function in the x/erc20 module keeper.
+// This function is called from the RPC client (eth_estimateGas) and internally.
 // When called from the RPC client, we need to reset the gas meter before
 // simulating the transaction to have
 // an accurate gas estimation for EVM extensions transactions.
@@ -311,8 +310,6 @@ func (k Keeper) EstimateGasInternal(c context.Context, req *types.EthCallRequest
 		}
 	}
 
-	// TODO: Recap the highest gas limit with account's available balance.
-
 	// Recap the highest gas allowance with specified gascap.
 	if req.GasCap != 0 && hi > req.GasCap {
 		hi = req.GasCap
@@ -334,6 +331,32 @@ func (k Keeper) EstimateGasInternal(c context.Context, req *types.EthCallRequest
 	msg, err := args.ToMessage(req.GasCap, cfg.BaseFee)
 	if err != nil {
 		return nil, status.Error(codes.Internal, err.Error())
+	}
+
+	// Recap the highest gas limit with account's available balance.
+	if msg.GasFeeCap().BitLen() != 0 {
+		balance := k.bankKeeper.GetBalance(ctx, sdk.AccAddress(args.From.Bytes()), k.GetParams(ctx).EvmDenom)
+		available := balance.Amount
+		transfer := "0"
+		if args.Value != nil {
+			if args.Value.ToInt().Cmp(available.BigInt()) >= 0 {
+				return nil, core.ErrInsufficientFundsForTransfer
+			}
+			available = available.Sub(sdkmath.NewIntFromBigInt(args.Value.ToInt()))
+			transfer = args.Value.String()
+		}
+		allowance := available.Quo(sdkmath.NewIntFromBigInt(msg.GasFeeCap()))
+
+		// If the allowance is larger than maximum uint64, skip checking
+		if allowance.IsUint64() && hi > allowance.Uint64() {
+			k.Logger(ctx).Debug("Gas estimation capped by limited funds", "original", hi, "balance", balance,
+				"sent", transfer, "maxFeePerGas", msg.GasFeeCap().String(), "fundable", allowance)
+
+			hi = allowance.Uint64()
+			if hi < ethparams.TxGas {
+				return nil, core.ErrInsufficientFunds
+			}
+		}
 	}
 
 	// NOTE: the errors from the executable below should be consistent with go-ethereum,